Move schema preprocessing to JSONSchemaForm component
